Add Exit command to list of commands of client
when the server stops the game or a game is finished or saved the client
has to stop and close it's socket connection with the server so resources
can be saved, in this times exit command will be issued from server and
there is the need for a handler in the client. ExitCommand will act like a handler
which in turn it make a new instace of the ClientCloser class that it can close all
connections.