Modify Command class to seperate command tokens corectly
in the command class a commnad string will be recived from the server
which has a predefined frame, the command handler in the client needs
to tokenize the string by the space and according to the last first
word in the command string call suitable command handler.
in the previous code the spliting was not done correctly and also
the was also a bug that if the command had lots of words separated by
a space it will only show the first letter as the message or the question
so fixed the problem.