Finish feature Gamestate known game loop state
therer is the need for the game state to keep track of the state of
the game loop, weather it is night or day or poll or introduction night,
to do so it use a GameLoopStateEnum and 4 accessor methods to change the
game loop state field.
this data is mainly used for saving game and auto loading game.